tidied up types
