+++ packages/ui-app/src/SideMenu/index.tsx
@@ -5,21 +5,12 @@ export {
 	CategoriesPopover,
 	type CategoriesPopoverProps,
 } from './Categories'
-export {
-	DefaultMenu,
-	type DefaultMenuBarItem,
-	type DefaultMenuProps,
-} from './DefaultMenu'
-export { FloatingMenu, type FloatingMenuProps } from './FloatingMenu'
-export { NavSimple, type NavSimpleProps } from './NavSimple'
+export { DefaultMenu } from './DefaultMenu'
+export { FloatingMenu } from './FloatingMenu'
+export { NavSimple } from './NavSimple'
 export { Primary } from './Primary'
 export type {
-	HeadingProps,
-	IconProps,
-	MinimisedProps,
-	PrimaryProps,
-	RenderSideMenuMain,
-	SecondaryProps,
+	DefaultMenuBarItem,
 	SideMenuMainRenderProps,
 } from './types'
 export {
